Python 3 support.
In Python 3.3, `inspect.trace()` returns an empty list which causes
another exception to be thrown. The source for `--pdb` in nose merely
calls `pdb.post_mortem(tb)`.  This function is, more or less,

    p = pdb.Pdb()
    p.reset()
    p.interaction(None, tb)

This seems to fix the bug. In addition, the frame where the exception
was thrown does seem to be found anyway as all of the local variables
are accessible.